@@ -4,11 +4,11 @@ version: 0.10.0
 homepage: https://github.com/Enough-Software/enough_convert
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 #dependencies:
 #  characters: ^1.0.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.4
+  pedantic: ^1.11.0
+  test: ^1.16.5
